requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cchardet==2.1.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
discord.py==1.6.0
Flask==1.1.2
html2text==2020.1.16
html5lib==1.0b10
htmlfetcher==0.0.6
htmlparsing==0.1.5
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
lassie==0.11.9
lxml==4.6.3
MarkupSafe==1.1.1
micawber==0.5.3
multidict==5.1.0
natsort==7.1.1
numpy==1.20.1
pandas==1.2.3
parse==1.19.0
path==15.1.2
Pillow==8.1.2
pyfiglet==0.8.post1
python-dateutil==2.8.1
python-oembed==0.2.4
pytz==2021.1
requests==2.25.1
robohash==1.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
toapi==2.1.2
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.6.3
